+++ packages/js-toolkit/Base/utils.ts
@@ -96,19 +96,13 @@ export function eventIsNative(event: string, element: HTMLElement): boolean {
 
 /**
  * Get the target of a given event.
- *
- * @todo Return false in v3 if event is not defined or not native to prevent adding a listener.
  */
 export function getEventTarget(
   instance: Base,
   event: string,
   config: BaseConfig,
-): Base | Base['$el'] {
-  if (eventIsDefinedInConfig(event, config)) {
-    return instance;
-  }
-
-  if (eventIsNative(event, instance.$el)) {
+): Base['$el'] {
+  if (eventIsDefinedInConfig(event, config) || eventIsNative(event, instance.$el)) {
     return instance.$el;
   }
 }